Remove scan-payment-card-detect
