+++ cardscan-ui/build.gradle
@@ -42,8 +42,7 @@ dependencies {
     api project(":scan-ui")
     api project(":scan-payment-base")
 
-    compileOnly project(":scan-payment-ocr")
-    compileOnly project(":scan-payment-card-detect")
+    compileOnly project(':scan-payment-ocr-full')
 
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.0"
     implementation "androidx.activity:activity-ktx:1.2.3"
@@ -60,8 +59,7 @@ dependencies {
     testImplementation "org.jetbrains.kotlin:kotlin-test:1.5.10"
     testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.5.0"
 
-    testImplementation project(":scan-payment-ocr")
-    testImplementation project(":scan-payment-card-detect")
+    testImplementation project(':scan-payment-ocr-full')
 }
 
 dependencies {
@@ -69,8 +67,7 @@ dependencies {
     androidTestImplementation "androidx.test.espresso:espresso-core:3.3.0"
     androidTestImplementation "org.jetbrains.kotlin:kotlin-test:1.5.10"
 
-    androidTestImplementation project(":scan-payment-ocr")
-    androidTestImplementation project(":scan-payment-card-detect")
+    androidTestImplementation project(':scan-payment-ocr-full')
 }
 
 apply from: 'deploy.gradle'
