--- scan-payment-card-detect/src/main/java/com/getbouncer/scan/payment/carddetect/CardDetectModelManager.kt
@@ -1,9 +1,0 @@
-package com.getbouncer.scan.payment.carddetect
-
-import android.content.Context
-import com.getbouncer.scan.framework.Fetcher
-import com.getbouncer.scan.payment.ModelManager
-
-object CardDetectModelManager : ModelManager() {
-    override fun getModelFetcher(context: Context): Fetcher = CardDetect.ModelFetcher(context)
-}
