+++ scan-payment-ocr-download/build.gradle
