+++ scan-payment-ocr-download/src/main/AndroidManifest.xml
