+++ scan-payment-ocr-full/.gitignore
