+++ scan-payment-ocr-full/build.gradle
