+++ scan-payment-ocr-full/src/androidTest/java/com/getbouncer/scan/payment/ocr/SSDOcrTest.kt
