+++ scan-payment-ocr-full/src/androidTest/res/drawable/card_no_pan.png
