+++ scan-payment-ocr-full/src/androidTest/res/drawable/card_pan.png
