+++ scan-payment-ocr-full/src/androidTest/res/drawable/ocr_card_numbers_clear.png
