+++ scan-payment-ocr-full/src/main/assets/darknite_1_1_1_16.tflite
