+++ scan-payment-ocr-full/src/main/assets/ux_0_5_23_16.tflite
