+++ scan-payment-ocr-minimal/src/androidTest/res/drawable/card_no_card.png
