+++ scan-payment-ocr-minimal/src/androidTest/res/drawable/card_pan.png
