+++ scan-payment-ocr-minimal/src/main/assets/UX.0.25.106.8.tflite
