--- scan-payment-ocr/.gitignore
@@ -1,1 +1,0 @@
-/build
