--- scan-payment-ocr/src/main/java/com/getbouncer/scan/payment/ocr/SSDOcrModelManager.kt
@@ -1,9 +1,0 @@
-package com.getbouncer.scan.payment.ocr
-
-import android.content.Context
-import com.getbouncer.scan.framework.Fetcher
-import com.getbouncer.scan.payment.ModelManager
-
-object SSDOcrModelManager : ModelManager() {
-    override fun getModelFetcher(context: Context): Fetcher = SSDOcr.ModelFetcher(context)
-}
