+++ settings.gradle
@@ -6,9 +6,8 @@ include ':scan-camera2'
 include ':scan-camerax'
 include ':scan-framework'
 include ':scan-payment-base'
-include ':scan-payment-card-detect'
-include ':scan-payment-card-detect-minimal'
-include ':scan-payment-ocr'
+include ':scan-payment-ocr-download'
+include ':scan-payment-ocr-full'
 include ':scan-payment-ocr-minimal'
 include ':scan-ui'
 include ':tensorflow-lite'
